#564 :recycle: separate BroadcastMessageToRead + BroadcastMessageToSend
