--- wls-gui-wahllokalsystem/src/api/wls-clients/broadcast-service/BroadcastMessage.ts
@@ -1,15 +1,0 @@
-export interface BroadcastMessageToRead {
-  readonly oid: string;
-  readonly wahlbezirkIDs: string[];
-  readonly nachricht: string;
-}
-
-export class BroadcastMessageToSend {
-  wahlbezirkIDs: string[];
-  nachricht: string;
-
-  constructor(wahlbezirkIDs: string[], nachricht: string) {
-    this.wahlbezirkIDs = wahlbezirkIDs;
-    this.nachricht = nachricht;
-  }
-}
