+++ wls-gui-wahllokalsystem/src/api/wls-clients/broadcast-service/BroadcastMessageToRead.ts
@@ -1,0 +1,5 @@
+export interface BroadcastMessageToRead {
+  readonly oid: string;
+  readonly wahlbezirkIDs: string[];
+  readonly nachricht: string;
+}
