+++ wls-gui-wahllokalsystem/src/api/wls-clients/broadcast-service/BroadcastMessageToSend.ts
@@ -1,0 +1,6 @@
+export default class BroadcastMessageToSend {
+  constructor(
+    public wahlbezirkIDs: string[],
+    public nachricht: string
+  ) {}
+}
